requirements.txt
asgiref==3.2.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==3.0.4
django-crispy-forms==1.9.0
django-filter==2.2.0
djangorestframework==3.11.0
ecdsa==0.15
idna==2.9
Markdown==3.2.1
oauthlib==3.1.0
#pkg-resources==0.0.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.12.0
python-jose==3.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.8
